requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.3
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.11.0
easyocr==1.7.1
ffmpeg-python==0.2.0
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.9.1
future==0.18.3
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.36
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.3
google-cloud-logging==3.6.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==20.1.0
html5lib==1.1
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
incremental==22.10.0
inquirer==3.1.3
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
litellm==0.1.690
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
networkx==3.1
ninja==1.11.1
numpy==1.26.0
open-interpreter==0.1.4
openai==0.27.10
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.0
pandas-datareader==0.10.0
parsel==1.8.1
pdfkit==1.0.0
Pillow==10.0.1
Protego==0.3.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
PyDispatcher==2.0.7
pydub==0.25.1
pydyf==0.7.0
Pygments==2.16.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyocr==0.8.5
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyphen==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
pytube==15.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
queuelib==1.6.2
readchar==4.0.5
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
rich==13.5.3
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.10.1
seaborn==0.12.2
service-identity==23.1.0
shapely==2.0.1
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.22.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.8.30
tiktoken==0.4.0
tinycss2==1.2.1
tldextract==3.5.0
tokentrim==0.1.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.2
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.3.7
wget==3.2
yarl==1.9.2
yfinance==0.2.28
zipp==3.16.2
zope.interface==6.0
zopfli==0.2.3